Bug fixing and working functions
Fixed a few errors which was caused when attempting to add images to an emails array. Created a function that stores the email arrays and images associated with the emails in a local storage. Local storage is then loaded on page load. Created a function that displays the images for the email selected in the drop down menu. Images now display correctly on the screen and display for the correct email. The user cannot enter the same email into the array if it is already present.