Button layout and image collection layout.
Applied layout changes to all buttons to be more user friendly for smaller devices. Buttons for smaller devices only will display none when at larger viewports. Images in a collection now display in the centre of the collection container and will display the maximum amount of images possible on a row without going outside the container. Implemented changes to reduce the amount of empty space within the collection display.